+++ lib/presentation/shell/app_shell.dart
@@ -10,6 +10,7 @@ import '../pages/search/search_page.dart';
 import '../providers/home_widget_providers.dart';
 import '../providers/selection_providers.dart';
 import '../widgets/library/offline_prefetcher.dart';
+import '../widgets/player/listening_recorder.dart';
 import '../widgets/player/mini_player.dart';
 import '../widgets/player/playback_persistor.dart';
 import 'glass_nav_bar.dart';
@@ -140,6 +141,7 @@ class _AppShellState extends ConsumerState<AppShell>
             _HomeWidgetBridge(),
             OfflinePrefetcher(),
             PlaybackPersistor(),
+            ListeningRecorder(),
             MiniPlayer(),
             GlassNavBar(),
           ],